Simplify memory handling in dimension.go with generics.
`tiledb_dimension_get_domain` does not copy the bytes from the domain
into the passed-in pointer; instead it uses the passed-in pointer as an
out value, which it writes to point to the memory that it owns. This
means that we can eliminate our allocating code. At the same time, we
also use generics to reduce a bunch of duplication.